pdf_preprocessor/requirements.txt
cmake==3.21.1.post1
path==16.2.0
Pillow==8.3.1
pkg_resources==0.0.0
python-poppler==0.2.2
